Move prob and marginals from dynamics to utils
